RF: Split file checking from iteration, allow checking individual files (#4)
`walk` only works on directories, but you might just want to run this on
a single file. This PR factors out file checking from the directory
walk. If `bids_dir` is actually a file, we just check it once.

This also moves the `FileNotFound` filter into a post-hoc filter, rather
than deep in the loop.